Add `NOTHING` action
This is a proposal for a change that allows you to leave the `action`
column blank to signify that you want the rule to match messages and
stop processing further rules, but not actually move the thread from
its current state (similar to the current `mark_important` column
behavior).

I'm having trouble correctly routing code review emails. I don't want CI
messages for other people's changes to move threads to my inbox, but if
I specify `archive` on those, and a CI message is processed in the same
run as a relevant message, it overrides the relevant message. I could
move the CI filter rule below, but that precludes a "catch-all" review
rule to route uncategorized reviews to my inbox.

If this seems OK, let me know and I can add tests and make it a nicer
thing.