Add parent labels
Given a label structure like:

- list
  - abc
  - def

If you wanted to search all of the emails in both `list-abc` and
`list-def`, you'd need to search `label:list-abc OR label:list-def`.
This is because Gmail label nesting is actually flat, not really
hierarchical.

This change updates the labeling such that parent labels are added as
well, so searching `label:list` would give all emails in both abc and
def lists.