Fix utils.ts imports (#59)
[ES imports are not supported by
Clasp](https://github.com/google/clasp/blob/master/docs/typescript.md#modules-exports-and-imports).
The work-around that we seem to have been relying on is that classes are
still global after the typescript build, so this simply puts the util
functions in a class as static functions for namespacing. This allows
the unit tests to still resolve imports, but the Clasp build works.